requirements.txt
apache-airflow==2.10.3
apache-airflow-providers-postgres==5.14.0
openpyxl==3.1.2
pandas==2.1.4
psycopg2-binary==2.9.10
xlrd==2.0.1
